+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 All notable changes to the "nba-ticker" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-- Initial release
+### Added
+- Status bar item to display scores
+- Command to update scores (no UI)